requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pyyaml==6.0.2

numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2

pytest==8.3.4
